// src/common/api/entities/sys/TypeRefs.ts
export type SystemKeysReturn = {
	_type: TypeRef<SystemKeysReturn>;
	_original?: SystemKeysReturn

	_format: NumberString;
	systemAdminPubRsaKey: null | Uint8Array;
	systemAdminPubKeyVersion: NumberString;
	freeGroupKey: Uint8Array;
	premiumGroupKey: Uint8Array;
	systemAdminPubEccKey: null | Uint8Array;
	systemAdminPubKyberKey: null | Uint8Array;
	freeGroupKeyVersion: NumberString;
	premiumGroupKeyVersion: NumberString;

	freeGroup: null | Id;
	premiumGroup: null | Id;
}
// ...
export type PublicKeyGetOut = {
	_type: TypeRef<PublicKeyGetOut>;
	_original?: PublicKeyGetOut

	_format: NumberString;
	pubRsaKey: null | Uint8Array;
	pubKeyVersion: NumberString;
	pubEccKey: null | Uint8Array;
	pubKyberKey: null | Uint8Array;

	signature: null | PublicKeySignature;
}
// ...
export type KeyMac = {
	_type: TypeRef<KeyMac>;
	_original?: KeyMac

	_id: Id;
	taggedKeyVersion: NumberString;
	tag: Uint8Array;
	taggingKeyVersion: NumberString;

	taggingGroup: Id;
}
// ...
export type PubDistributionKey = {
	_type: TypeRef<PubDistributionKey>;
	_original?: PubDistributionKey

	_id: Id;
	pubEccKey: Uint8Array;
	pubKyberKey: Uint8Array;

	userGroupId: Id;
	pubKeyMac: KeyMac;
}
// ...
export type PublicKeySignature = {
	_type: TypeRef<PublicKeySignature>;
	_original?: PublicKeySignature

	_id: Id;
	signature: Uint8Array;
	signingKeyVersion: NumberString;
	signatureType: NumberString;
	publicKeyVersion: NumberString;
}